[IMP] add model changes for test
The exercises for this chapter ask users to add some code to the core
code to test against. Since we probably want this solution to work
without any changes (otherwise it wouldn't be a solution?), let's add
in that missing code so the tests all pass out of the box.